Moved BlockPairWrapper
